moving config from events/yaml to core/config
