+++ server/core/config/raw/pre_workflow_step.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 
 	validation "github.com/go-ozzo/ozzo-validation"
-	"github.com/runatlantis/atlantis/server/events/yaml/valid"
+	"github.com/runatlantis/atlantis/server/core/config/valid"
 )
 
 // PreWorkflowHook represents a single action/command to perform. In YAML,
