+++ server/core/config/raw/pre_workflow_step_test.go
@@ -3,8 +3,8 @@ package raw_test
 import (
 	"testing"
 
-	"github.com/runatlantis/atlantis/server/events/yaml/raw"
-	"github.com/runatlantis/atlantis/server/events/yaml/valid"
+	"github.com/runatlantis/atlantis/server/core/config/raw"
+	"github.com/runatlantis/atlantis/server/core/config/valid"
 	. "github.com/runatlantis/atlantis/testing"
 	yaml "gopkg.in/yaml.v2"
 )
