+++ server/core/config/valid/global_cfg.go
