+++ server/core/config/valid/global_cfg_test.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/hashicorp/go-version"
 	"github.com/mohae/deepcopy"
-	"github.com/runatlantis/atlantis/server/events/yaml"
-	"github.com/runatlantis/atlantis/server/events/yaml/valid"
+	"github.com/runatlantis/atlantis/server/core/config"
+	"github.com/runatlantis/atlantis/server/core/config/valid"
 	"github.com/runatlantis/atlantis/server/logging"
 	. "github.com/runatlantis/atlantis/testing"
 )
