+++ server/core/config/valid/global_repo_cfg.go
