+++ server/core/runtime/policy/conftest_client_test.go
@@ -15,7 +15,7 @@ import (
 	conftest_mocks "github.com/runatlantis/atlantis/server/core/runtime/policy/mocks"
 	terraform_mocks "github.com/runatlantis/atlantis/server/core/terraform/mocks"
 	"github.com/runatlantis/atlantis/server/events/models"
-	"github.com/runatlantis/atlantis/server/events/yaml/valid"
+	"github.com/runatlantis/atlantis/server/core/config/valid"
 	"github.com/runatlantis/atlantis/server/logging"
 	. "github.com/runatlantis/atlantis/testing"
 )
