+++ server/core/runtime/policy/mocks/matchers/valid_policyset.go
@@ -5,7 +5,7 @@ import (
 	"github.com/petergtz/pegomock"
 	"reflect"
 
-	valid "github.com/runatlantis/atlantis/server/events/yaml/valid"
+	valid "github.com/runatlantis/atlantis/server/core/config/valid"
 )
 
 func AnyValidPolicySet() valid.PolicySet {
