+++ server/events/models/models.go
@@ -30,7 +30,7 @@ import (
 	"github.com/uber-go/tally"
 
 	"github.com/pkg/errors"
-	"github.com/runatlantis/atlantis/server/events/yaml/valid"
+	"github.com/runatlantis/atlantis/server/core/config/valid"
 )
 
 const (
